refactor: Remove repository cache migrations
