+++ lib/util/cache/repository/impl/base.ts
@@ -4,23 +4,11 @@ import is from '@sindresorhus/is';
 import hasha from 'hasha';
 import { GlobalConfig } from '../../../../config/global';
 import { logger } from '../../../../logger';
+import { match } from '../../../schema';
 import { safeStringify } from '../../../stringify';
-import {
-  CACHE_REVISION,
-  isValidRev10,
-  isValidRev11,
-  isValidRev12,
-  isValidRev13,
-} from '../common';
-import type {
-  RepoCache,
-  RepoCacheData,
-  RepoCacheRecord,
-  RepoCacheRecordV10,
-  RepoCacheRecordV11,
-  RepoCacheRecordV12,
-  RepoCacheRecordV13,
-} from '../types';
+import { CACHE_REVISION } from '../common';
+import { RepoCacheRecord, RepoCacheV13 } from '../schemas';
+import type { RepoCache, RepoCacheData } from '../types';
 
 const compress = promisify(zlib.brotliCompress);
 const decompress = promisify(zlib.brotliDecompress);
@@ -39,30 +27,16 @@ export abstract class RepoCacheBase implements RepoCache {
 
   protected abstract write(data: RepoCacheRecord): Promise<void>;
 
-  private restoreFromRev10(oldCache: RepoCacheRecordV10): void {
-    delete oldCache.repository;
-    delete oldCache.revision;
-    this.data = oldCache;
-  }
-
-  private restoreFromRev11(oldCache: RepoCacheRecordV11): void {
-    this.data = oldCache.data;
-  }
-
-  private async restoreFromRev12(oldCache: RepoCacheRecordV12): Promise<void> {
+  private async restore(oldCache: RepoCacheRecord): Promise<void> {
+    if (oldCache.fingerprint !== this.fingerprint) {
+      logger.debug('Repository cache fingerprint is invalid');
+      return;
+    }
     const compressed = Buffer.from(oldCache.payload, 'base64');
     const uncompressed = await decompress(compressed);
     const jsonStr = uncompressed.toString('utf8');
     this.data = JSON.parse(jsonStr);
     this.oldHash = oldCache.hash;
-  }
-
-  private async restoreFromRev13(oldCache: RepoCacheRecordV13): Promise<void> {
-    if (oldCache.fingerprint !== this.fingerprint) {
-      logger.debug('Repository cache fingerprint is invalid');
-      return;
-    }
-    await this.restoreFromRev12(oldCache);
   }
 
   async load(): Promise<void> {
@@ -76,27 +50,9 @@ export abstract class RepoCacheBase implements RepoCache {
       }
       const oldCache = JSON.parse(rawOldCache) as unknown;
 
-      if (isValidRev13(oldCache)) {
-        await this.restoreFromRev13(oldCache);
+      if (match(RepoCacheV13, oldCache, false)) {
+        await this.restore(oldCache);
         logger.debug('Repository cache is restored from revision 13');
-        return;
-      }
-
-      if (isValidRev12(oldCache, this.repository)) {
-        await this.restoreFromRev12(oldCache);
-        logger.debug('Repository cache is restored from revision 12');
-        return;
-      }
-
-      if (isValidRev11(oldCache, this.repository)) {
-        this.restoreFromRev11(oldCache);
-        logger.debug('Repository cache is restored from revision 11');
-        return;
-      }
-
-      if (isValidRev10(oldCache, this.repository)) {
-        this.restoreFromRev10(oldCache);
-        logger.debug('Repository cache is restored from revision 10');
         return;
       }
