+++ lib/util/cache/repository/types.ts
@@ -75,30 +75,6 @@ export interface RepoCacheData {
   prComments?: Record<number, Record<string, string>>;
 }
 
-export interface RepoCacheRecordV10 extends RepoCacheData {
-  repository?: string;
-  revision?: number;
-}
-
-export interface RepoCacheRecordV11 {
-  repository: string;
-  revision: number;
-  data: RepoCacheData;
-}
-
-export interface RepoCacheRecordV12 {
-  repository: string;
-  revision: number;
-  payload: string;
-  hash: string;
-}
-
-export interface RepoCacheRecordV13 extends RepoCacheRecordV12 {
-  fingerprint: string;
-}
-
-export type RepoCacheRecord = RepoCacheRecordV13;
-
 export interface RepoCache {
   load(): Promise<void>;
   save(): Promise<void>;
